Update 3_Bit_Asynchronous_Up_Counter_Design.v
The 3-bit Asynchronous Up Counter contains three J-K flip-flops and the J-K input of all the flip-flops are connected to ‘1’. All these flip-flops are negative edge triggered but the outputs change asynchronously. The clock signal is directly applied to the first J-K flip-flop. So, the output of first J-K flip-flop toggles for every negative edge of clock signal.