Create 3_Bit_Asynchronous_Down_Counter_Testbench.v
The 3-bit Asynchronous Up Counter contains 3 T flip-flops and the T-input of all the flip-flops are connected to ‘1’. All these flip-flops are negative edge triggered but the outputs change asynchronously. The clock signal is directly applied to the first T flip-flop. So, the output of first T flip-flop toggles for every negative edge of clock signal.