Create MOD_6_Counter_Testbench.v
MOD-6 Counter

A Mod-6 counter is a digital circuit that counts from 0 to 5 (or 6 states) and then repeats. The counter produces six unique states, hence the name "Mod-6".

The counter can be implemented using various flip-flops, such as D flip-flops, JK flip-flops, or T flip-flops. One possible implementation is to use three JK flip-flops and connect their outputs to a logic circuit that detects the count of 5 and resets the counter to 0.

The Mod-6 counter can be useful in various applications, such as in digital clocks or in systems that require cyclic behavior. It can also be cascaded with other counters to create larger counters for more complex applications.