Create 3to1_MUX_Testbench.v
3to1 Multiplexer

A 3 to 1 multiplexer is a digital circuit that has three inputs and one output. It selects one of the three inputs based on a select signal, and sends that input to the output. In other words, it is like a switch that connects one of the three inputs to the output based on the value of the select signal. The selected input is then propagated to the output of the multiplexer.