Create Clock_Divider_Testbench.v
Clock Divider

A clock divider is a digital circuit that reduces the frequency of a clock signal by a certain factor. The output frequency of the divided clock is a fraction of the input clock frequency, where the fraction is determined by the division factor. Clock dividers are commonly used in digital systems to create slower clock signals for timing purposes, or to synchronize different parts of a system that require different clock frequencies. In Verilog, we can implement clock dividers using simple counter logic to toggle an output signal at a slower frequency than the input clock signal.