Create SISO Design.v
SERIAL IN SERIAL OUT(SISO) SHIFT REGISTER

A serial-in, serial-out shift register is a type of digital circuit that can store and shift a stream of binary data serially, one bit at a time. It is commonly used in digital systems for data transmission, storage, and synchronization.

The basic structure of a serial-in, serial-out shift register consists of a series of flip-flops connected in a chain, with the output of each flip-flop connected to the input of the next. The input data is fed into the first flip-flop, and the output data is read out from the last flip-flop.

The shift register operates by shifting the data from one flip-flop to the next on each clock cycle. This is typically controlled by a clock signal that triggers the flip-flops to change state. On each clock cycle, the input data is latched into the first flip-flop, and the data in each flip-flop is shifted to the next flip-flop in the chain. The output data can be read out from the last flip-flop in the chain.