Create Waveform.md
Parallel In Serial Out(PISO) Shift Register

A parallel-in, serial-out (PISO) shift register is a type of digital circuit that can store parallel data and shift it out serially, one bit at a time. It is commonly used in digital systems for interfacing with parallel data sources, such as memory or input/output devices, and for serializing the data for transmission over a serial communication link.

The basic structure of a PISO shift register is similar to a serial-in, serial-out (SISO) shift register, with a chain of flip-flops connected in series. However, the PISO shift register also includes a parallel input data bus that allows multiple bits of data to be loaded into the shift register simultaneously.