Update 4. Waveform.md
Free Running Shift Register

A free-running shift register is a digital circuit that consists of a chain of flip-flops connected in series, where the output of each flip-flop is connected to the input of the next one. The circuit operates by shifting the data through the flip-flops, with each flip-flop storing one bit of data.

In a free-running shift register, the input data is continuously shifted through the flip-flops in a cyclic pattern, with the output of the last flip-flop being fed back to the input of the first flip-flop. This creates a loop that allows the shift register to continue shifting the data indefinitely, hence the name "free-running."

The speed at which the data is shifted through the flip-flops is determined by the clock frequency. Each time the clock signal transitions from low to high (or vice versa), the shift register advances by one bit. The clock frequency can be adjusted to control the speed of the shift register.

Free-running shift registers are commonly used in digital signal processing (DSP) applications, such as audio and video processing, where data needs to be shifted through a circuit at a constant rate. They can also be used in applications that require serial-to-parallel or parallel-to-serial data conversion.